requirements.txt
pillow~=7.0.0
pyshp~=2.1.0
